+++ src/page/list/controller.ctrl.js
@@ -1,0 +1,6 @@
+export default class ListController {
+	namespace = 'list';
+	showList(params, state) {
+		return params;
+	}
+}